requirements.txt
boto3==1.26.139
botocore==1.29.139
Flask==3.0.3
Flask-Cors==4.0.0
Jinja2==3.1.2
MarkupSafe==2.1.1
python-dotenv==1.0.0
s3transfer==0.6.0
urllib3==1.26.15
Werkzeug==3.0.3